episode_001/requirements.txt
# Python dependencies for Episode 001 - World GDP Shift

pandas>=2.0.0
PyYAML>=6.0
requests>=2.31.0
tqdm>=4.65.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
